Added a few types, refactored some mtg stuff.

+++ src/kaflib/applications/mtg/CollectorPanel.java
@@ -14,7 +14,10 @@ import java.text.NumberFormat;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
@@ -30,11 +33,14 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
 import kaflib.graphics.GraphicsUtils;
+import kaflib.gui.ImageComponent;
 import kaflib.gui.ProgressLabel;
 import kaflib.types.WordTree;
 import kaflib.types.Worker;
 import kaflib.utils.FileUtils;
+import kaflib.utils.GUIUtils;
 import kaflib.utils.RandomUtils;
+import kaflib.utils.StringUtils;
 
 /**
  * Defines a web collector with a coupled UI (I know, right?).
@@ -49,21 +55,39 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 	private final JFormattedTextField max;
 	private final JFormattedTextField value;
 	private final JCheckBox tor_only;
+	private final JButton file_browse;
 	private final JButton start;
 	private final ProgressLabel progress;
 	private final JPanel center;
 	
+	private int max_valid;
+	
+	private static final String CSS_NAME_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_NAME_END = "nameRow .value";
+	private static final String CSS_MANA_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_MANA_END = "manaRow .value";
+	private static final String CSS_CMC_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_CMC_END = "cmcRow .value";
+	private static final String CSS_TYPES_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_TYPES_END = "typeRow .value";
+	private static final String CSS_TEXT_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_TEXT_END = "textRow .value .cardtextbox";
+	private static final String CSS_PT_START = "#ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_";
+	private static final String CSS_PT_END = "ptRow .value";
+	
 	private Boolean tor_available;
-	public static final int MAX_ID = 438850;
 	private final BufferedImage no_image;
+	private final ImageComponent image;
 	private final CardDatabase db;
 	
 	public CollectorPanel(final CardDatabase db, final WordTree tree) throws Exception {	
-		super(new BorderLayout());
+		super(new FlowLayout());
 		this.db = db;
 		tor_available = null;
 		self = this;
 		this.tree = tree;
+		
+		max_valid = db.getMaxValid();
 		
 		NumberFormat format = NumberFormat.getInstance();
 		format.setGroupingUsed(false);
@@ -78,7 +102,7 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 		min.addKeyListener(this);
 		max = new JFormattedTextField(formatter);
 		max.setColumns(6);
-		max.setValue(MAX_ID);
+		max.setValue(max_valid);
 		max.addKeyListener(this);
 		value = new JFormattedTextField(formatter);
 		value.setColumns(6);
@@ -92,33 +116,50 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 		
 		no_image = db.getDefaultImage();
 		
-		center = new JPanel(new GridLayout(2, 3));
-		JPanel panel = new JPanel(new FlowLayout());
-
+		center = new JPanel(new BorderLayout());
+		
+		JPanel panel = GUIUtils.getTitledPanel("Random collect");
+		panel.setLayout(new FlowLayout());
 		panel.add(new JLabel("Count:"));
 		panel.add(count);
-		center.add(panel);
-		panel = new JPanel(new FlowLayout());
 		panel.add(new JLabel("from:"));
 		panel.add(min);
-		center.add(panel);
-		panel = new JPanel(new FlowLayout());
 		panel.add(new JLabel("to:"));
 		panel.add(max);
-		center.add(panel);
+		center.add(panel, BorderLayout.NORTH);
 
-		panel = new JPanel(new FlowLayout());
-		panel.add(new JLabel("Card ID:"));
-		panel.add(value);
-		center.add(panel);
+		panel = GUIUtils.getTitledPanel("List");
+		panel.setLayout(new BorderLayout());
+		file_browse = new JButton("Browse...");
+		file_browse.addActionListener(this);
+		panel.add(file_browse, BorderLayout.WEST);
+		center.add(panel, BorderLayout.CENTER);
 		
-		panel = new JPanel(new FlowLayout());
-		center.add(panel);
-		center.add(start);
+		panel = GUIUtils.getTitledPanel("One-off");
+		panel.setLayout(new BorderLayout());
+		panel.setAlignmentX(LEFT_ALIGNMENT);
+		JLabel label = new JLabel("Card ID:");
+		panel.add(label, BorderLayout.WEST);
+		panel.add(value, BorderLayout.CENTER);
 		
-		add(tor_only, BorderLayout.NORTH);
-		add(center, BorderLayout.CENTER);
-		add(progress, BorderLayout.SOUTH);
+		center.add(panel, BorderLayout.SOUTH);
+
+		panel = new JPanel(new BorderLayout());
+		panel.add(progress, BorderLayout.CENTER);
+		panel.add(start, BorderLayout.EAST);
+		
+		image = new ImageComponent();
+		image.update(no_image);
+		
+		JPanel left_panel = new JPanel();
+		left_panel.setLayout(new BorderLayout());
+		
+		left_panel.add(tor_only, BorderLayout.NORTH);
+		left_panel.add(center, BorderLayout.CENTER);
+		left_panel.add(panel, BorderLayout.SOUTH);
+		add(left_panel);
+		add(image, BorderLayout.EAST);
+
 	}
 	
 	private static int randomNeighbor(final int seed, 
@@ -198,14 +239,27 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
  		return values;
 	}
 	
+
+	private void scrape() throws Exception {
+		scrape(null);
+	}
+	
 	/**
 	 * Perform scrape based on inputs.  Do not call from ui thread.
 	 * @param count
 	 * @throws Exception
 	 */
-	private void scrape() throws Exception {
+	private void scrape(final File file) throws Exception {
 		Collection<Integer> values;
-		if (!value.getText().isEmpty()) {
+		if (file != null) {
+			values = new HashSet<Integer>();
+			for (String line : FileUtils.readLines(file)) {
+				if (StringUtils.isInt(line.trim())) { 
+					values.add(Integer.valueOf(line.trim()));
+				}
+			}
+		}
+		else if (!value.getText().isEmpty()) {
 			values = new ArrayList<Integer>();
 			values.add(Integer.valueOf(value.getText()));
 			value.setText("");
@@ -214,11 +268,11 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 		else if (!count.getText().isEmpty() &&
 				 !min.getText().isEmpty() &&
 				 !max.getText().isEmpty()) {
-			values = generateHaveTargets(Integer.valueOf(count.getText()));
+//			values = generateHaveTargets(Integer.valueOf(count.getText()));
 
-//			values = generateTargets(Integer.valueOf(count.getText()), 
-//								  Integer.valueOf(min.getText()), 
-//								  Integer.valueOf(max.getText()));
+			values = generateTargets(Integer.valueOf(count.getText()), 
+								  Integer.valueOf(min.getText()), 
+								  Integer.valueOf(max.getText()));
 		}
 		else {
 			JOptionPane.showMessageDialog(this, "Must specify value or count/min/max.");
@@ -258,95 +312,165 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 		return temp.replaceAll("\"\\s+align.*\\>", "]");
 	}
 	
-	private Card parseCard(final String html, final int id) throws Exception {
+	
+	
+	private Set<CardInstance> parseCards(final String html, final int id) throws Exception {
 		Document document = Jsoup.parse(html);
-		Elements elements;
+		Set<CardInstance> set = new HashSet<CardInstance>();
 		
+		Elements elements = document.select(".cardDetails");
+		if (elements.size() == 1) {
+			set.add(parse(elements.first(), id, ""));			
+		}
+		else if (elements.size() == 2) { 
+			set.add(parse(elements.get(0), id, "ctl02_"));	
+			set.add(parse(elements.get(1), id, "ctl03_"));	
+		}
+		else {
+		}
+		
+		return set;
+	}
+		
+
+
+	/**
+	 * Search for id matches in the document.
+	 * @param document
+	 * @return
+	 * @throws Exception
+	 */
+	private Integer parseID(final Element document) throws Exception {
+		Pattern pattern = Pattern.compile("^(\\d+)\\D.*", Pattern.MULTILINE);
+		String chunks[] = document.outerHtml().split("multiverseid\\=");
+		for (int i = 1; i < chunks.length; i++) {
+			String chunk = StringUtils.truncateIf(chunks[i], 16);
+			Matcher matcher = pattern.matcher(chunk);
+			if (matcher.matches()) {
+				return Integer.valueOf(matcher.group(1));
+			}			
+		}
+		return null;
+	}
+	
+	
+	private CardInstance parse(final Element document, final int id, final String substr) throws Exception {
+		Elements elements;
 		String name = "";
 		String cost = "";
 		String cmc = "";
 		String types = "";
 		String text = "";
 		String pt = "";
-		
+		try {
+			Integer found_id = parseID(document);
+			if (found_id == null) {
+				found_id = id;
+			}
+			
+			elements = document.select(CSS_NAME_START + substr + CSS_NAME_END);
+			if (elements == null || elements.size() != 1) {
+				System.err.println(CSS_NAME_START + substr + CSS_NAME_END);
+				if (elements != null) {
+					System.err.println("Matches " + elements.size() + " elements.");
+				}
+				throw new Exception("Could not parse name.");
+			}
+			if (elements.size() == 1) {
+				name = elements.get(0).text();
+			}
+			
 
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_nameRow .value");
-		if (elements.size() != 1) {
-			return null;
-		}
-		if (elements.size() == 1) {
-			name = elements.get(0).text();
-		}
-		
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_manaRow .value");
-		for (Element element : elements) {
-			cost = cost + replaceColors(element.html()) + "\n";
-		}
-		
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_cmcRow .value");
-		if (elements.size() > 1) {
-			throw new Exception("Multiple cmc: " + html);
-		}
-		if (elements.size() == 1) {
-			cmc = elements.get(0).text();
-		}
-		
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_typeRow .value");
-		for (Element element : elements) {
-			types = types + element.text() + "\n";
-		}
-		
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_textRow .value .cardtextbox");
-		for (Element element : elements) {
-			text = text + replaceColors(element.html()) + "\n";
-			text = text.replace("<i>", "");
-			text = text.replace("</i>", "");
-		}
-		
-		elements = document.select(".cardDetails #ctl00_ctl00_ctl00_MainContent_SubContent_SubContent_ptRow .value");
-		if (elements.size() > 1) {
-			throw new Exception("Multiple p/t: " + html);
-		}
-		if (elements.size() == 1) {
-			pt = elements.get(0).text();
-		}
-		
-		Card card = new Card(
-							id,
-							name,
-							cost,
-							cmc,
-							types,
-							text,
-							pt);
+			elements = document.select(CSS_MANA_START + substr + CSS_MANA_END);
+			if (elements == null) {
+				throw new Exception("Could not parse cost.");
+			}
+			for (Element element : elements) {
+				cost = cost + replaceColors(element.html()) + "\n";
+			}
 
-		
-		float rating = 0;
-		int votes = 0;
-		
-		elements = document.select(".textRatingValue");
-		if (elements.size() > 1) {
-			throw new Exception("Multiple rating tags: " + html);
-		}
-		if (elements.size() == 1) {
-			rating = Float.valueOf(elements.get(0).text());
-		}
+			elements = document.select(CSS_CMC_START + substr + CSS_CMC_END);
+			if (elements.size() > 1) {
+				throw new Exception("Multiple cmc: " + found_id + ".");
+			}
+			if (elements.size() == 1) {
+				cmc = elements.get(0).text();
+			}
+			
 
-		elements = document.select(".totalVotesValue");
-		if (elements.size() > 1) {
-			throw new Exception("Multiple votes tags: " + html);
-		}
-		if (elements.size() == 1) {
-			votes = Integer.valueOf(elements.get(0).text());
-		}
+			elements = document.select(CSS_TYPES_START + substr + CSS_TYPES_END);
+			if (elements == null) {
+				throw new Exception("Could not parse types.");
+			}
+			for (Element element : elements) {
+				types = types + element.text() + "\n";
+			}
+			
 
-		if (votes >= 15) {
-			card.setCommunityRating(rating);
-			card.setCommunityVotes(votes);
+			elements = document.select(CSS_TEXT_START + substr + CSS_TEXT_END);
+			if (elements == null) {
+				throw new Exception("Could not parse text.");
+			}
+			for (Element element : elements) {
+				text = text + replaceColors(element.html()) + "\n";
+				text = text.replace("<i>", "");
+				text = text.replace("</i>", "");
+			}
+			
+
+			elements = document.select(CSS_PT_START + substr + CSS_PT_END);
+			if (elements.size() > 1) {
+				throw new Exception("Multiple p/t: " + found_id + ".");
+			}
+			if (elements.size() == 1) {
+				pt = elements.get(0).text();
+			}
+			
+			CardInstance card = new CardInstance(
+								found_id,
+								name,
+								cost,
+								cmc,
+								types,
+								text,
+								pt);
+	
+			
+			float rating = 0;
+			int votes = 0;
+			
+			elements = document.select(".textRatingValue");
+			if (elements != null) {
+				if (elements.size() > 1) {
+					throw new Exception("Multiple rating tags: " + found_id + ".");
+				}
+				if (elements.size() == 1) {
+					rating = Float.valueOf(elements.get(0).text());
+				}
+			}
+	
+			elements = document.select(".totalVotesValue");
+			if (elements != null) {
+				if (elements.size() > 1) {
+					throw new Exception("Multiple votes tags: " + found_id + ".");
+				}
+				if (elements.size() == 1) {
+					votes = Integer.valueOf(elements.get(0).text());
+				}
+			}
+	
+			if (votes >= 15) {
+				card.setCommunityRating(rating);
+				card.setCommunityVotes(votes);
+			}
+			return card;
+		}
+		catch (Exception e) {
+			System.err.println("Error parsing " + id + ".");
+			System.err.println("Html: " + document.outerHtml());
+			throw e;
 		}
 		
-		
-		return card;
 	}
 	
 	private boolean torAvailable() {
@@ -414,18 +538,23 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 			db.addForeign(id);
 		}
 		else {
-			Card card = parseCard(get(new URL("http://gatherer.wizards.com/Pages/Card/Details.aspx?printed=false&multiverseid=" + id)), id);
+			Set<CardInstance> cards = parseCards(get(new URL("http://gatherer.wizards.com/Pages/Card/Details.aspx?printed=false&multiverseid=" + id)), id);
 			
-			if (card != null) {
-				db.add(card);
-				tree.insert(card.getName());
-				
-				File image_file = new File(db.getImageDirectory(), id + ".png");
-				get(image_file, new URL("http://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=" + id + "&type=card"));
-	
-				BufferedImage image = GraphicsUtils.read(image_file);
-				if (GraphicsUtils.equal(no_image, image)) {
-					image_file.delete();
+			if (cards != null) {
+				for (CardInstance card : cards) {
+					if (card != null) {
+						db.add(card);
+						tree.insert(card.getName());
+						File image_file = new File(db.getImageDirectory(), card.getIDString() + ".png");
+						get(image_file, new URL("http://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=" + card.getIDString() + "&type=card"));
+						BufferedImage image = GraphicsUtils.read(image_file);
+						if (GraphicsUtils.equal(no_image, image)) {
+							image_file.delete();
+						}
+						else {
+							this.image.update(image);
+						}
+					}
 				}
 			}
 			else {
@@ -457,6 +586,32 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 				};
 				worker.start();
 			}
+			else if (e.getSource().equals(file_browse)) {
+				final File file = GUIUtils.chooseFile(this, db.getImageDirectory().getParentFile());
+				if (file == null) {
+					return;
+				}
+				start.setEnabled(false);
+				Worker worker = new Worker() {
+					@Override
+					protected void process() throws Exception {
+						try {
+							scrape(file);
+						}
+						catch (Exception e) {
+							JOptionPane.showMessageDialog(self, "Scrape error:\n" + e.getMessage());
+						}
+						finally {
+							start.setEnabled(true);
+						}
+					}
+					
+				};
+				worker.start();
+			}
+			else {
+				
+			}
 		}
 		catch (Exception ex) {
 			JOptionPane.showMessageDialog(self, "Could not handle action.");
@@ -473,7 +628,7 @@ public class CollectorPanel extends JPanel implements KeyListener, ActionListene
 		else {
 			count.setText("");
 			min.setText("1");
-			max.setText("" + MAX_ID);
+			max.setText("" + max_valid);
 		}
 	}
